requirements.txt
absl-py==1.3.0
aiosignal==1.3.1
asttokens==2.0.5
astunparse==1.6.3
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
brotlipy==0.7.0
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.4
colorama==0.4.4
colorlog==4.7.2
conda==22.9.0
conda-build==3.22.0
conda-content-trust==0+unknown
conda-package-handling==1.8.1
cryptography==36.0.0
decorator==5.1.1
distlib==0.3.6
exceptiongroup==1.0.0
executing==0.8.3
expecttest==0.1.4
filelock==3.6.0
frozenlist==1.3.3
future==0.18.2
glob2==0.7
google-auth==2.14.1
google-auth-oauthlib==0.4.6
grpcio==1.50.0
hypothesis==6.56.4
idna==3.3
importlib-metadata==5.0.0
ipython==8.4.0
jedi==0.18.1
Jinja2==2.11.3
joblib==1.2.0
jsonschema==4.17.0
libarchive-c==2.9
line-profiler==4.0.3
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
mpmath==1.2.1
msgpack==1.0.4
numpy==1.22.3
oauthlib==3.2.2
pandas==1.5.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pip==23.0.1
pkginfo==1.8.3
platformdirs==2.5.4
plotly==5.11.0
prompt-toolkit==3.0.20
protobuf==3.20.3
psutil==5.8.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.3
pycparser==2.21
Pygments==2.11.2
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyrsistent==0.19.2
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
ray==2.1.0
recbole==1.1.1
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.9
ruamel-yaml-conda==0.15.100
scikit-learn==1.1.3
scipy==1.9.3
setuptools==61.2.0
six==1.16.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
stack-data==0.2.0
sympy==1.11.1
tabulate==0.9.0
tenacity==8.1.0
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.11.2
torch==1.13.0
torch-geometric==2.1.0.post1
torch-scatter==2.0.9
torch-sparse==0.6.15
torchtext==0.14.0
torchvision==0.14.0
tqdm==4.63.0
traitlets==5.1.1
types-dataclasses==0.6.6
typing_extensions==4.4.0
urllib3==1.26.8
virtualenv==20.16.7
wcwidth==0.2.5
Werkzeug==2.2.2
wheel==0.37.1
zipp==3.10.0
